# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI React Backend",
    "version": "1.0.0"
  },
  "paths": {
    "/api/blog/": {
      "get": {
        "tags": [
          "blog"
        ],
        "summary": "Get All Blog Posts",
        "description": "\u3059\u3079\u3066\u306e\u30d6\u30ed\u30b0\u8a18\u4e8b\u3092\u53d6\u5f97\u3059\u308b",
        "operationId": "get_all_blog_posts_api_blog__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BlogPost"
                  },
                  "type": "array",
                  "title": "Response Get All Blog Posts Api Blog  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "blog"
        ],
        "summary": "Create Blog Post",
        "description": "\u65b0\u3057\u3044\u30d6\u30ed\u30b0\u8a18\u4e8b\u3092\u4f5c\u6210\u3059\u308b",
        "operationId": "create_blog_post_api_blog__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogPostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/blog/{post_id}": {
      "get": {
        "tags": [
          "blog"
        ],
        "summary": "Get Blog Post",
        "description": "\u6307\u5b9a\u3055\u308c\u305fID\u306e\u30d6\u30ed\u30b0\u8a18\u4e8b\u3092\u53d6\u5f97\u3059\u308b",
        "operationId": "get_blog_post_api_blog__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "blog"
        ],
        "summary": "Update Blog Post",
        "description": "\u65e2\u5b58\u306e\u30d6\u30ed\u30b0\u8a18\u4e8b\u3092\u66f4\u65b0\u3059\u308b",
        "operationId": "update_blog_post_api_blog__post_id__put",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogPostUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "blog"
        ],
        "summary": "Delete Blog Post",
        "description": "\u30d6\u30ed\u30b0\u8a18\u4e8b\u3092\u524a\u9664\u3059\u308b",
        "operationId": "delete_blog_post_api_blog__post_id__delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Post Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "health_check_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/{full_path}": {
      "get": {
        "summary": "Serve React App",
        "operationId": "serve_react_app__full_path__get",
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "parameters": [
          {
            "name": "full_path",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Full Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BlogPost": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "author",
          "id",
          "created_at"
        ],
        "title": "BlogPost",
        "description": "\u30d6\u30ed\u30b0\u8a18\u4e8b\u30ec\u30b9\u30dd\u30f3\u30b9\u30e2\u30c7\u30eb"
      },
      "BlogPostCreate": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "author": {
            "type": "string",
            "title": "Author"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "author"
        ],
        "title": "BlogPostCreate",
        "description": "\u30d6\u30ed\u30b0\u8a18\u4e8b\u4f5c\u6210\u30e2\u30c7\u30eb"
      },
      "BlogPostUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author"
          }
        },
        "type": "object",
        "title": "BlogPostUpdate",
        "description": "\u30d6\u30ed\u30b0\u8a18\u4e8b\u66f4\u65b0\u30e2\u30c7\u30eb\uff08\u3059\u3079\u3066\u306e\u30d5\u30a3\u30fc\u30eb\u30c9\u306f\u7701\u7565\u53ef\u80fd\uff09"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBasic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}